Refactor, create selection page, standardize constants
