+++ App/SelectionPage/SelectionPage.java
@@ -1,0 +1,16 @@
+package App.SelectionPage;
+
+// Imports
+import java.awt.*;
+import java.awt.event.*;
+import javax.swing.*;
+
+import  App.Constants.SelectionConstants;
+
+public class SelectionPage extends JPanel {
+
+    // Constructor
+    public SelectionPage(){
+
+    }
+}
